Refactor library organization
