+++ src/ban/mod.rs
@@ -1,220 +1,40 @@
-use crate::LintLevel;
+mod cfg;
+mod graph;
+
+use self::cfg::TreeSkip;
+use crate::{Diagnostic, LintLevel};
 use anyhow::{Context, Error};
-use codespan_reporting::diagnostic::Diagnostic;
-use rayon::prelude::*;
 use semver::{Version, VersionReq};
-use serde::Deserialize;
-use std::{cmp, collections::HashMap, fmt};
+use std::cmp::Ordering;
 
-#[derive(Deserialize)]
-pub struct CrateId {
-    // The name of the crate
-    pub name: String,
-    /// The version constraints of the crate
-    #[serde(default = "any")]
-    pub version: VersionReq,
-}
-
-#[derive(Deserialize)]
-pub struct TreeSkip {
-    #[serde(flatten)]
-    pub id: CrateId,
-    pub depth: Option<usize>,
-}
-
-const fn lint_warn() -> LintLevel {
-    LintLevel::Warn
-}
-
-fn any() -> VersionReq {
-    VersionReq::any()
-}
-
-const fn highlight() -> GraphHighlight {
-    GraphHighlight::All
-}
-
-#[derive(Deserialize, PartialEq, Eq, Copy, Clone)]
-#[serde(rename_all = "kebab-case")]
-pub enum GraphHighlight {
-    /// Highlights the path to a duplicate dependency with the fewest number
-    /// of total edges, which tends to make it the best candidate for removing
-    SimplestPath,
-    /// Highlights the path to the duplicate dependency with the lowest version
-    LowestVersion,
-    /// Highlights with all of the other configs
-    All,
-}
-
-impl GraphHighlight {
-    #[inline]
-    fn simplest(self) -> bool {
-        self == Self::SimplestPath || self == Self::All
-    }
-
-    #[inline]
-    fn lowest_version(self) -> bool {
-        self == Self::LowestVersion || self == Self::All
-    }
-}
-
-#[derive(Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub struct Config {
-    /// Disallow multiple versions of the same crate
-    #[serde(default = "lint_warn")]
-    pub multiple_versions: LintLevel,
-    /// How the duplicate graphs are highlighted
-    #[serde(default = "highlight")]
-    pub highlight: GraphHighlight,
-    /// The crates that will cause us to emit failures
-    #[serde(default)]
-    pub deny: Vec<toml::Spanned<CrateId>>,
-    /// If specified, means only the listed crates are allowed
-    #[serde(default)]
-    pub allow: Vec<toml::Spanned<CrateId>>,
-    /// If specified, disregards the crate completely
-    #[serde(default)]
-    pub skip: Vec<toml::Spanned<CrateId>>,
-    /// If specified, disregards the crate's transitive dependencies
-    /// down to a certain depth
-    #[serde(default)]
-    pub skip_tree: Vec<toml::Spanned<TreeSkip>>,
-}
-
-impl Config {
-    pub fn validate(
-        self,
-        cfg_file: codespan::FileId,
-        _contents: &str,
-    ) -> Result<ValidConfig, Vec<codespan_reporting::diagnostic::Diagnostic>> {
-        use codespan_reporting::diagnostic::Label;
-
-        let from = |s: toml::Spanned<CrateId>| {
-            let span = s.start() as u32..s.end() as u32;
-            let inner = s.into_inner();
-            KrateId {
-                name: inner.name,
-                version: inner.version,
-                span,
-            }
-        };
-
-        let mut diagnostics = Vec::new();
-
-        let mut denied: Vec<_> = self.deny.into_iter().map(from).collect();
-        denied.par_sort();
-
-        let mut allowed: Vec<_> = self.allow.into_iter().map(from).collect();
-        allowed.par_sort();
-
-        let mut skipped: Vec<_> = self.skip.into_iter().map(from).collect();
-        skipped.par_sort();
-
-        let mut add_diag = |first: (&KrateId, &str), second: (&KrateId, &str)| {
-            let flabel = Label::new(
-                cfg_file,
-                first.0.span.clone(),
-                format!("marked as `{}`", first.1),
-            );
-            let slabel = Label::new(
-                cfg_file,
-                second.0.span.clone(),
-                format!("marked as `{}`", second.1),
-            );
-
-            // Put the one that occurs last as the primary label to make it clear
-            // that the first one was "ok" until we noticed this other one
-            let diag = if flabel.span.start() > slabel.span.start() {
-                Diagnostic::new_error(
-                    format!(
-                        "a license id was specified in both `{}` and `{}`",
-                        first.1, second.1
-                    ),
-                    flabel,
-                )
-                .with_secondary_labels(std::iter::once(slabel))
-            } else {
-                Diagnostic::new_error(
-                    format!(
-                        "a license id was specified in both `{}` and `{}`",
-                        second.1, first.1
-                    ),
-                    slabel,
-                )
-                .with_secondary_labels(std::iter::once(flabel))
-            };
-
-            diagnostics.push(diag);
-        };
-
-        for d in &denied {
-            if let Ok(ai) = allowed.binary_search(&d) {
-                add_diag((d, "deny"), (&allowed[ai], "allow"));
-            }
-            if let Ok(si) = skipped.binary_search(&d) {
-                add_diag((d, "deny"), (&skipped[si], "skip"));
-            }
-        }
-
-        for a in &allowed {
-            if let Ok(si) = skipped.binary_search(&a) {
-                add_diag((a, "allow"), (&skipped[si], "skip"));
-            }
-        }
-
-        if !diagnostics.is_empty() {
-            Err(diagnostics)
-        } else {
-            Ok(ValidConfig {
-                file_id: cfg_file,
-                multiple_versions: self.multiple_versions,
-                highlight: self.highlight,
-                denied,
-                allowed,
-                skipped,
-                tree_skipped: self.skip_tree,
-            })
-        }
-    }
-}
+pub use self::cfg::{Config, ValidConfig};
 
 #[derive(Eq)]
-struct KrateId {
+pub struct KrateId {
     name: String,
     version: VersionReq,
     span: std::ops::Range<u32>,
 }
 
 impl Ord for KrateId {
-    fn cmp(&self, o: &Self) -> cmp::Ordering {
+    fn cmp(&self, o: &Self) -> Ordering {
         match self.name.cmp(&o.name) {
-            cmp::Ordering::Equal => self.version.cmp(&o.version),
+            Ordering::Equal => self.version.cmp(&o.version),
             o => o,
         }
     }
 }
 
 impl PartialOrd for KrateId {
-    fn partial_cmp(&self, o: &Self) -> Option<cmp::Ordering> {
+    fn partial_cmp(&self, o: &Self) -> Option<Ordering> {
         Some(self.cmp(o))
     }
 }
 
 impl PartialEq for KrateId {
     fn eq(&self, o: &Self) -> bool {
-        self.cmp(o) == cmp::Ordering::Equal
+        self.cmp(o) == Ordering::Equal
     }
-}
-
-pub struct ValidConfig {
-    pub file_id: codespan::FileId,
-    pub multiple_versions: LintLevel,
-    pub highlight: GraphHighlight,
-    denied: Vec<KrateId>,
-    allowed: Vec<KrateId>,
-    skipped: Vec<KrateId>,
-    tree_skipped: Vec<toml::Spanned<TreeSkip>>,
 }
 
 fn binary_search<'a>(
@@ -224,7 +44,7 @@ fn binary_search<'a>(
     let lowest = VersionReq::exact(&Version::new(0, 0, 0));
 
     match arr.binary_search_by(|i| match i.name.cmp(&details.name) {
-        cmp::Ordering::Equal => i.version.cmp(&lowest),
+        Ordering::Equal => i.version.cmp(&lowest),
         o => o,
     }) {
         Ok(i) => Ok((i, &arr[i])),
@@ -259,7 +79,7 @@ fn binary_search_by_name<'a>(
     let lowest = Version::new(0, 0, 0);
 
     match arr.binary_search_by(|i| match i.name.as_str().cmp(name) {
-        cmp::Ordering::Equal => i.version.cmp(&lowest),
+        Ordering::Equal => i.version.cmp(&lowest),
         o => o,
     }) {
         Ok(i) | Err(i) => {
@@ -280,445 +100,6 @@ fn binary_search_by_name<'a>(
             Ok(begin..end)
         }
     }
-}
-
-use petgraph::Graph;
-
-#[derive(Hash, Copy, Clone, PartialEq, Eq)]
-struct Node<'a> {
-    name: &'a str,
-    version: &'a Version,
-}
-
-impl<'a, 'b: 'a> From<&'b crate::KrateDetails> for Node<'a> {
-    fn from(d: &'b crate::KrateDetails) -> Self {
-        Self {
-            name: &d.name,
-            version: &d.version,
-        }
-    }
-}
-
-impl<'a> fmt::Debug for Node<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}({})", self.name, self.version)
-    }
-}
-
-impl<'a> fmt::Display for Node<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}({})", self.name, self.version)
-    }
-}
-
-type Id = petgraph::graph::NodeIndex<u32>;
-
-fn append_dependency_chain<'a>(
-    crates: &'a crate::Krates,
-    start: usize,
-    graph: &mut Graph<Node<'a>, &'a str>,
-    node_map: &mut HashMap<Node<'a>, Id>,
-) {
-    let cd = &crates.krates[start];
-
-    let root_node = Node::from(cd);
-    let root_id = graph.add_node(root_node);
-
-    node_map.insert(root_node, root_id);
-
-    let mut node_stack = vec![(root_node, root_id)];
-
-    while let Some(node) = node_stack.pop() {
-        let parents: Vec<_> = crates
-            .as_ref()
-            .par_iter()
-            .filter_map(|cd| {
-                crates
-                    .resolved
-                    .nodes
-                    .binary_search_by(|rp| rp.id.cmp(&cd.id))
-                    .ok()
-                    .and_then(|i| {
-                        crates.resolved.nodes[i]
-                            .dependencies
-                            .binary_search_by(|did| {
-                                let mut iter = did.repr.splitn(3, char::is_whitespace);
-                                match iter.next() {
-                                    Some(n) => match n.cmp(&node.0.name) {
-                                        cmp::Ordering::Equal => iter
-                                            .next()
-                                            .and_then(|version| {
-                                                version
-                                                    .parse::<Version>()
-                                                    .ok()
-                                                    .map(|v| v.cmp(node.0.version))
-                                            })
-                                            .unwrap_or(cmp::Ordering::Less),
-                                        o => o,
-                                    },
-                                    None => cmp::Ordering::Less,
-                                }
-                            })
-                            .ok()
-                            .and_then(|_| crates.crate_by_id(&crates.resolved.nodes[i].id))
-                    })
-            })
-            .collect();
-
-        for parent in parents.into_iter().map(Node::from) {
-            match node_map.get(&parent) {
-                Some(id) => {
-                    if !graph.contains_edge(*id, node.1) {
-                        graph.add_edge(*id, node.1, "");
-                    }
-                }
-                None => {
-                    let id = graph.add_node(parent);
-
-                    node_map.insert(parent, id);
-                    graph.add_edge(id, node.1, "");
-
-                    node_stack.push((parent, id));
-                }
-            }
-        }
-    }
-}
-
-#[allow(non_camel_case_types)]
-#[derive(Debug)]
-pub enum Shape {
-    r#box,
-    polygon,
-    ellipse,
-    oval,
-    circle,
-    point,
-    egg,
-    triangle,
-    plaintext,
-    plain,
-    diamond,
-    trapezium,
-    parallelogram,
-    house,
-    pentagon,
-    hexagon,
-    septagon,
-    octagon,
-    doublecircle,
-    doubleoctagon,
-    tripleoctagon,
-    invtriangle,
-    invtrapezium,
-    invhouse,
-    Mdiamond,
-    Msquare,
-    Mcircle,
-    rect,
-    rectangle,
-    square,
-    star,
-    none,
-    underline,
-    cylinder,
-    note,
-    tab,
-    folder,
-    box3d,
-    component,
-    promoter,
-    cds,
-    terminator,
-    utr,
-    primersite,
-    restrictionsite,
-    fivepoverhang,
-    threepoverhang,
-    noverhang,
-    assembly,
-    signature,
-    insulator,
-    ribosite,
-    rnastab,
-    proteasesite,
-    proteinstab,
-    rpromoter,
-    rarrow,
-    larrow,
-    lpromoter,
-}
-
-#[allow(non_camel_case_types)]
-#[derive(Debug)]
-pub enum Style {
-    filled,
-    invisible,
-    diagonals,
-    rounded,
-    dashed,
-    dotted,
-    solid,
-    bold,
-}
-
-struct NodeAttributes<'a> {
-    label: Option<&'a dyn fmt::Display>,
-    shape: Option<Shape>,
-    style: Option<Style>,
-    color: Option<&'static str>,
-    fill_color: Option<&'static str>,
-}
-
-impl<'a> Default for NodeAttributes<'a> {
-    fn default() -> Self {
-        Self {
-            label: None,
-            shape: None,
-            style: None,
-            color: None,
-            fill_color: None,
-        }
-    }
-}
-
-impl<'a> NodeAttributes<'a> {
-    fn has_attrs(&self) -> bool {
-        self.label.is_some()
-            || self.shape.is_some()
-            || self.style.is_some()
-            || self.color.is_some()
-            || self.fill_color.is_some()
-    }
-}
-
-#[derive(Default)]
-struct EdgeAttributes {
-    color: Option<&'static str>,
-}
-
-const INDENT: &str = "    ";
-
-fn create_graph(
-    dup_name: &str,
-    highlight: GraphHighlight,
-    krates: &crate::Krates,
-    duplicates: &[usize],
-) -> Result<String, Error> {
-    use petgraph::visit::{EdgeRef, NodeRef};
-
-    let mut graph = Graph::new();
-    let mut node_map = HashMap::new();
-
-    for duplicate in duplicates {
-        append_dependency_chain(krates, *duplicate, &mut graph, &mut node_map);
-    }
-
-    let mut edges = Vec::with_capacity(duplicates.len());
-    let mut dupes: HashMap<&str, Vec<_>> = HashMap::new();
-
-    for duplicate in duplicates {
-        let dest = Node::from(&krates.krates[*duplicate]);
-
-        if let Some(id) = node_map.get(&dest) {
-            let mut nodes = vec![*id];
-            let mut set = std::collections::HashSet::new();
-            while let Some(id) = nodes.pop() {
-                let node = graph.node_weight(id).unwrap();
-                dupes
-                    .entry(node.name)
-                    .and_modify(|v| {
-                        if !v.contains(&node.version) {
-                            v.push(node.version);
-                        }
-                    })
-                    .or_insert_with(|| vec![node.version]);
-
-                for node in graph.neighbors_directed(id, petgraph::Direction::Incoming) {
-                    set.insert(graph.find_edge(node, id).unwrap());
-                    nodes.push(node);
-                }
-            }
-
-            edges.push(set);
-        }
-    }
-
-    dupes.retain(|_, v| {
-        v.sort();
-        v.len() > 1
-    });
-
-    // Find the version with the least number of total edges to the least common ancestor,
-    // this will presumably be the easiest version to "fix"
-    // This just returns the first lowest one, there can be multiple with
-    // same number of edges
-    let smollest = edges
-        .iter()
-        .min_by(|a, b| a.len().cmp(&b.len()))
-        .context("expected shortest edge path")?;
-    let lowest = &edges[0];
-
-    print_graph(
-        &graph,
-        |node| {
-            let node_weight = node.weight();
-
-            if node_weight.name == dup_name || dupes.contains_key(node_weight.name) {
-                NodeAttributes {
-                    label: Some(node_weight.version),
-                    shape: Some(Shape::r#box),
-                    color: Some("red"),
-                    style: Some(Style::rounded),
-                    ..Default::default()
-                }
-            } else {
-                NodeAttributes {
-                    label: Some(node.1),
-                    shape: Some(Shape::r#box),
-                    style: Some(Style::rounded),
-                    ..Default::default()
-                }
-            }
-        },
-        |edge| {
-            if highlight.simplest() && smollest.contains(&edge.id()) {
-                EdgeAttributes { color: Some("red") }
-            } else if highlight.lowest_version() && lowest.contains(&edge.id()) {
-                EdgeAttributes {
-                    color: Some("blue"),
-                }
-            } else {
-                EdgeAttributes { color: None }
-            }
-        },
-        |output| {
-            use std::fmt::Write;
-
-            for (i, (name, versions)) in dupes.iter().enumerate() {
-                writeln!(output, "{}subgraph cluster_{} {{", INDENT, i)?;
-
-                write!(output, "{}{}{{rank=same ", INDENT, INDENT)?;
-
-                for version in versions {
-                    if let Some(id) = node_map.get(&Node { name, version }) {
-                        write!(output, "{} ", id.index())?;
-                    }
-                }
-
-                writeln!(
-                    output,
-                    "}}\n{}{}style=\"rounded{}\";\n{}{}label=\"{}\"\n{}}}",
-                    INDENT,
-                    INDENT,
-                    if name == &dup_name { ",filled" } else { "" },
-                    INDENT,
-                    INDENT,
-                    name,
-                    INDENT
-                )?;
-            }
-
-            Ok(())
-        },
-    )
-}
-
-fn print_graph<'a: 'b, 'b, NP, EP, SG>(
-    graph: &'a Graph<Node<'a>, &'a str>,
-    node_print: NP,
-    edge_print: EP,
-    subgraphs: SG,
-) -> Result<String, Error>
-where
-    NP: Fn((Id, &'b Node<'a>)) -> NodeAttributes<'b>,
-    EP: Fn(&petgraph::graph::EdgeReference<'_, &'a str, u32>) -> EdgeAttributes,
-    SG: Fn(&mut String) -> Result<(), Error>,
-{
-    use petgraph::visit::{EdgeRef, IntoNodeReferences, NodeIndexable, NodeRef};
-    use std::fmt::Write;
-    let mut output = String::with_capacity(1024);
-    writeln!(output, "digraph {{",)?;
-
-    // output all nodes
-    for node in graph.node_references() {
-        write!(output, "{}{}", INDENT, graph.to_index(node.id()))?;
-
-        let attrs = node_print(node);
-
-        if !attrs.has_attrs() {
-            writeln!(output)?;
-            continue;
-        }
-
-        let mut append = false;
-        write!(output, " [")?;
-
-        if let Some(label) = attrs.label {
-            write!(output, "label=\"{}\"", label)?;
-            append = true;
-        }
-
-        if let Some(shape) = attrs.shape {
-            write!(
-                output,
-                "{}shape={:?}",
-                if append { ", " } else { "" },
-                shape
-            )?;
-            append = true;
-        }
-
-        if let Some(style) = attrs.style {
-            write!(
-                output,
-                "{}style={:?}",
-                if append { ", " } else { "" },
-                style
-            )?;
-            append = true;
-        }
-
-        if let Some(color) = attrs.color {
-            write!(output, "{}color={}", if append { ", " } else { "" }, color)?;
-            append = true;
-        }
-
-        if let Some(color) = attrs.fill_color {
-            write!(
-                output,
-                "{}fillcolor={}",
-                if append { ", " } else { "" },
-                color
-            )?;
-        }
-
-        writeln!(output, "]")?;
-    }
-
-    // output all edges
-    for edge in graph.edge_references() {
-        write!(
-            output,
-            "{}{} -> {}",
-            INDENT,
-            graph.to_index(edge.source()),
-            graph.to_index(edge.target()),
-        )?;
-
-        let attrs = edge_print(&edge);
-
-        if let Some(color) = attrs.color {
-            writeln!(output, " [color={}]", color)?;
-        } else {
-            writeln!(output)?;
-        }
-    }
-
-    subgraphs(&mut output)?;
-
-    writeln!(output, "}}")?;
-    Ok(output)
 }
 
 pub type Pid = cargo_metadata::PackageId;
@@ -776,7 +157,7 @@ pub struct DupGraph {
     pub graph: String,
 }
 
-pub fn check_bans<OG>(
+pub fn check<OG>(
     krates: &crate::Krates,
     cfg: ValidConfig,
     (lock_id, lock_contents): (codespan::FileId, &str),
@@ -1019,7 +400,7 @@ where
                     }
 
                     if let Some(ref og) = output_graph {
-                        let graph = create_graph(
+                        let graph = graph::create_graph(
                             multi_detector.name,
                             cfg.highlight,
                             krates,
@@ -1069,6 +450,7 @@ where
 
 #[cfg(test)]
 mod test {
+    use super::cfg::CrateId;
     use super::*;
 
     #[test]
